refactor(server/utils): move utility files into utils module
